Install.rst: Adding pipenv support
Pipenv support is added for coala installation for Linux, Mac and Windows.
However, Pipfile is not supported natively yet.
It helps manage pip and virtualenv together.
Pipenv generated Pipfile and Pipfile.lock for it's functioning and will
be updated by pipenv itself on every install and update of packages,
this is the reason these 2 files are needed to be ignored inside
.gitignore.

Closes https://github.com/coala/documentation/issues/502